fixed tosca typo
