+++ plugins/tosca/src/main/resources/META-INF/de.unistuttgart.iaas.amyassist.amy-plugin-systemtime.natlangMeta
@@ -1,0 +1,2 @@
+.aims:
+ToscaSpeech.aim.xml
