+++ plugins/tosca/src/main/resources/META-INF/de.unistuttgart.iaas.amyassist.amy-plugin-tosca.properties
