+++ plugins/tosca/src/main/resources/META-INF/de.unistuttgart.iaas.amyassist.amy.core.di.Services
@@ -1,0 +1,1 @@
+de.unistuttgart.iaas.amyassist.amy.plugin.tosca.ToscaLogic
