--- plugins/toska/src/main/resources/META-INF/de.unistuttgart.iaas.amyassist.amy-plugin-systemtime.natlangMeta
@@ -1,2 +1,0 @@
-.aims:
-ToskaSpeech.aim.xml
