--- plugins/toska/src/main/resources/META-INF/de.unistuttgart.iaas.amyassist.amy.core.console.Console
@@ -1,1 +1,0 @@
-de.unistuttgart.iaas.amyassist.amy.plugin.toska.ToskaConsole
