+++ pom.xml
@@ -74,7 +74,7 @@
 		<module>plugins/systemtime</module>
 		<module>plugins/weather</module>
 		<module>plugins/navigation</module>
-		<module>plugins/toska</module>
+		<module>plugins/tosca</module>
 		<module>amy-message-hub-api</module>
 		<module>amy-message-hub</module>
 		<module>remote-sr</module>
